Updated user_user_group table migration.

+++ src/database/migrations/2014_03_19_185043_create_shibboleth_user_user_group_table.php
@@ -13,15 +13,16 @@ class CreateShibbolethUserUserGroupTable extends Migration
      */
     public function up()
     {
-        Schema::create('user_user_group', function (Blueprint $table) {
-            $table->increments('id');
+        if(!Schema::hasTable('user_user_group')) {
 
-            $table->unsignedInteger('user_group_id');
-            $table->unsignedInteger('user_id');
-
-            $table->foreign('user_group_id')->references('id')->on('user_groups')->onDelete('cascade');
-            $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
-        });
+            Schema::create('user_user_group', function (Blueprint $table) {
+                $table->integer('user_group_id')->unsigned()->index();
+                $table->foreign('user_group_id')->references('id')->on('user_groups')->onDelete('cascade');
+                $table->integer('user_id')->unsigned()->index();
+                $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
+                $table->primary(['user_group_id', 'user_id']);
+            });
+        }
 
     }
